+++ thinglinks-modules/thinglinks-modules-link/src/main/java/com/mqttsnet/thinglinks/link/common/asyncthread/LinkAsyncConfig.java
@@ -1,4 +1,4 @@
-package com.mqttsnet.thinglinks.tdengine.common.asyncthread;
+package com.mqttsnet.thinglinks.link.common.asyncthread;
 
 import com.mqttsnet.thinglinks.common.core.asyncthread.SuRejectHandle;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,12 +12,12 @@ import java.util.concurrent.Executor;
 
 /**
  * 异步@Async多线程配置
- *
+ * 使用原则：每一个业务场景使用独立的线程池，不要使用全局共享线程池
  * @author thinglinks
  */
 @Configuration
 @EnableAsync
-public class AsyncConfig {
+public class LinkAsyncConfig {
 
     /**
      * 核心线程数（默认线程数）
@@ -48,8 +48,10 @@ public class AsyncConfig {
     @Autowired
     private SuRejectHandle rejectHandle;
 
-
-    @Bean("tdengineAsync")
+    /**
+     * link服务全局共享异步线程池
+     */
+    @Bean("linkAsync")
     public Executor tdengineAsync() {
         ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
         executor.setCorePoolSize(corePoolSize);//核心线程数
@@ -58,7 +60,7 @@ public class AsyncConfig {
         //其他传参使用不缓存SynchronousQueue
         executor.setQueueCapacity(queueCapacity);//队列长度(超过队列长度无法存储,则开启最大线程数)
         executor.setKeepAliveSeconds(keepAliveTime);//空闲线程最大存活时间 默认60s
-        executor.setThreadNamePrefix(threadNamePrefix+"tdengineAsync-");//线程名前缀
+        executor.setThreadNamePrefix(threadNamePrefix+"linkAsync-");//线程名前缀
         executor.setRejectedExecutionHandler(rejectHandle);// 自定义任务丢失处理策略   该策略输出由scheduling-1打印
         //设置线程池等待所有任务都完成再关闭
         executor.setWaitForTasksToCompleteOnShutdown(true);
@@ -66,4 +68,27 @@ public class AsyncConfig {
         executor.initialize();
         return executor;
     }
+
+    /**
+     * MQTT设备消息消费异步线程池配置
+     */
+    @Bean("linkAsync-mqttMsg")
+    public Executor linkAsync() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setCorePoolSize(corePoolSize);//核心线程数
+        executor.setMaxPoolSize(maxPoolSize);//最大线程数    cpu核数/(1-0.8)//cup核数*2//cup核数+1
+        //传参正值使用无界LinkedBlockingQueue
+        //其他传参使用不缓存SynchronousQueue
+        executor.setQueueCapacity(queueCapacity);//队列长度(超过队列长度无法存储,则开启最大线程数)
+        executor.setKeepAliveSeconds(keepAliveTime);//空闲线程最大存活时间 默认60s
+        executor.setThreadNamePrefix(threadNamePrefix+"linkAsync-mqttMsg");//线程名前缀
+        executor.setRejectedExecutionHandler(rejectHandle);//自定义任务丢失处理策略   该策略输出由scheduling-1打印
+        executor.setWaitForTasksToCompleteOnShutdown(true);
+        executor.setAwaitTerminationSeconds(keepAliveTime);
+        executor.initialize();
+        return executor;
+    }
+
+
+
 }
