+++ thinglinks-modules/thinglinks-modules-tdengine/src/main/java/com/mqttsnet/thinglinks/tdengine/common/asyncthread/TdengineAsyncConfig.java
@@ -1,4 +1,4 @@
-package com.mqttsnet.thinglinks.rule.common.asyncthread;
+package com.mqttsnet.thinglinks.tdengine.common.asyncthread;
 
 import com.mqttsnet.thinglinks.common.core.asyncthread.SuRejectHandle;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +17,7 @@ import java.util.concurrent.Executor;
  */
 @Configuration
 @EnableAsync
-public class AsyncConfig {
+public class TdengineAsyncConfig {
 
     /**
      * 核心线程数（默认线程数）
@@ -49,8 +49,8 @@ public class AsyncConfig {
     private SuRejectHandle rejectHandle;
 
 
-    @Bean("ruleAsync")
-    public Executor ruleAsync() {
+    @Bean("tdengineAsync")
+    public Executor tdengineAsync() {
         ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
         executor.setCorePoolSize(corePoolSize);//核心线程数
         executor.setMaxPoolSize(maxPoolSize);//最大线程数  cpu核数/(1-0.8)//cup核数*2//cup核数+1
@@ -58,7 +58,7 @@ public class AsyncConfig {
         //其他传参使用不缓存SynchronousQueue
         executor.setQueueCapacity(queueCapacity);//队列长度(超过队列长度无法存储,则开启最大线程数)
         executor.setKeepAliveSeconds(keepAliveTime);//空闲线程最大存活时间 默认60s
-        executor.setThreadNamePrefix(threadNamePrefix+"ruleAsync-");//线程名前缀
+        executor.setThreadNamePrefix(threadNamePrefix+"tdengineAsync-");//线程名前缀
         executor.setRejectedExecutionHandler(rejectHandle);// 自定义任务丢失处理策略   该策略输出由scheduling-1打印
         //设置线程池等待所有任务都完成再关闭
         executor.setWaitForTasksToCompleteOnShutdown(true);
